Fix: Make td-idf a double division

+++ src/scoring/tf_idf.cpp
@@ -7,7 +7,8 @@ TfIdf::TfIdf(uint32_t doc_count) : doc_count(doc_count) {}
 double TfIdf::score(const DocStats& doc_stats, const QueryStats& query_stats) const {
   double result = 0;
   for (const auto& word : query_stats.query_words) {
-    result += (word.frequency / doc_stats.doc_length) * idf(doc_count, word.total_count);
+    result += (static_cast<double>(word.frequency) / static_cast<double>(doc_stats.doc_length)) *
+              idf(doc_count, word.total_count);
   }
   return result;
 }
